refactoring | #47 | delete Import component
- Combined the components responsible for working with the file system into one
- Remove Import component
