+++ packages/cli-ui/src/pages/index.ts
@@ -1,7 +1,6 @@
 export { default as Home } from './Home'
-export { default as Import } from './Import'
 export { default as Projects } from './Projects'
 export { default as PageNotFound } from './PageNotFound'
 export { default as Depend } from './Depend'
-export { default as SelectCreateFolder } from './SelectCreateFolder'
+export { default as SelectFolder } from './SelectFolder'
 export { default as CreateProject } from './CreateProject'
